requirements.txt
aiogram==3.10.0
supabase==2.5.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
